dashboard/requirements.txt
streamlit==1.32.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
plotly==5.18.0
bcrypt==3.2.0
pyjwt==2.8.0
uuid==1.30 